+++ gems/aws-sdk-core/lib/aws-sdk-core/assume_role_credentials.rb
@@ -20,7 +20,7 @@ module Aws
   class AssumeRoleCredentials
 
     include CredentialProvider
-    include RefreshingCredentials
+    include ResilientRefreshingCredentials
 
     # @option options [required, String] :role_arn
     # @option options [required, String] :role_session_name
@@ -49,7 +49,6 @@ module Aws
         end
       end
       @client = client_opts[:client] || STS::Client.new(client_opts)
-      @async_refresh = true
       @metrics = ['CREDENTIALS_STS_ASSUME_ROLE']
       super
     end
@@ -60,7 +59,25 @@ module Aws
     # @return [Hash]
     attr_reader :assume_role_params
 
+    # STS error codes that indicate a misconfiguration (bad policy, denied
+    # access, disabled region, etc). Retrying will not resolve them, so they
+    # are raised immediately rather than backed off.
+    # @api private
+    NON_RECOVERABLE_ERROR_CODES = %w[
+      AccessDenied
+      IDPRejectedClaim
+      InvalidIdentityToken
+      MalformedPolicyDocument
+      PackedPolicyTooLarge
+      RegionDisabled
+    ].freeze
+
     private
+
+    def non_recoverable_error?(error)
+      error.is_a?(Aws::Errors::ServiceError) &&
+        NON_RECOVERABLE_ERROR_CODES.include?(error.code)
+    end
 
     def refresh
       resp = @client.assume_role(@assume_role_params)